+++ Resources/Locale/ru-RU/ss220/clothing/beer_cap.ftl
@@ -1,4 +1,4 @@
 ent-ClothingHeadBeerCap = Пивная кепка
     .desc = Кепка с двумя пивными банками. Налейте любимый напиток внутрь и утоляйте жажду на ходу.
-ent-ActionDrinkCap = Сделать глоток
+ent-ActionBeerCap = Сделать глоток
     .desc = Сделать глоток напитка, хранящегося в кепке.
