+++ Resources/Locale/ru-RU/ss220/recipes/tags.ftl
@@ -191,3 +191,7 @@ construction-graph-tag-Contrposter = Недопустимый постер
 construction-graph-tag-Legalposter = Допустимый постер
 
 construction-graph-tag-butter-slice = ломтик сливочного масла
+
+construction-graph-tag-red-cap = красная кепка
+construction-graph-tag-cablecuffs = самодельные наручники
+construction-graph-tag-beer-can = пустая пивная банка
