+++ Resources/Prototypes/SS220/Entities/Clothing/Head/beer_cap.yml
@@ -50,9 +50,12 @@
         type: TransferAmountBoundUserInterface
   - type: ActionGrant
     actions:
-    - ActionDrinkCap
+    - ActionBeerCap
   - type: ItemActionGrant
     actions:
-    - ActionDrinkCap
+    - ActionBeerCap
     activeIfWorn: true
+  - type: Construction
+    graph: BeerCapGraph
+    node: beer_cap
